refactor folder structure for react features
